Moved new claims to SubmitClaimStore
