+++ Projects/App/Sources/Journeys/ClaimsJourney.swift
@@ -68,15 +68,15 @@ extension AppJourney {
 
     private static func honestyPledge(from origin: ClaimsOrigin) -> some JourneyPresentation {
         HostingJourney(
-            ClaimsStore.self,
+            SubmitClaimStore.self,
             rootView: LoadingViewWithContent(.startClaim) {
                 HonestyPledge {
                     let ugglanStore: UgglanStore = globalPresentableStoreContainer.get()
                     if ugglanStore.state.askForPushNotificationPermission() {
-                        let store: ClaimsStore = globalPresentableStoreContainer.get()
+                        let store: SubmitClaimStore = globalPresentableStoreContainer.get()
                         store.send(.navigationAction(action: .openNotificationsPermissionScreen))
                     } else {
-                        let store: ClaimsStore = globalPresentableStoreContainer.get()
+                        let store: SubmitClaimStore = globalPresentableStoreContainer.get()
                         store.send(.startClaimRequest(with: origin.id))
                     }
                 }
@@ -91,10 +91,10 @@ extension AppJourney {
             if case let .navigationAction(navigationAction) = action {
                 if case .openNotificationsPermissionScreen = navigationAction {
                     HostingJourney(
-                        ClaimsStore.self,
+                        SubmitClaimStore.self,
                         rootView: LoadingViewWithContent(.startClaim) {
                             ClaimFlowAskForPushnotifications(onActionExecuted: {
-                                let store: ClaimsStore = globalPresentableStoreContainer.get()
+                                let store: SubmitClaimStore = globalPresentableStoreContainer.get()
                                 store.send(.startClaimRequest(with: origin.id))
                             })
                         },
