+++ Projects/Claims/Sources/Views/SubmitClaim/AudioRecording/SubmitClaimAudioRecordingScreen.swift
@@ -8,7 +8,7 @@ import hCoreUI
 import hGraphQL
 
 public struct SubmitClaimAudioRecordingScreen: View {
-    @PresentableStore var store: ClaimsStore
+    @PresentableStore var store: SubmitClaimStore
     @ObservedObject var audioPlayer: AudioPlayer
     @ObservedObject var audioRecorder: AudioRecorder
 
@@ -29,7 +29,7 @@ public struct SubmitClaimAudioRecordingScreen: View {
         LoadingViewWithContent(.postAudioRecording) {
             hForm {
                 PresentableStoreLens(
-                    ClaimsStore.self,
+                    SubmitClaimStore.self,
                     getter: { state in
                         state.audioRecordingStep
                     }
